parses raw all data

+++ ultimateData/ultimateData/src/main/java/data/raw/RawData.java
@@ -274,11 +274,11 @@ public class RawData {
 		return component2.durationPointEstimate;
 	}
 
-	public Short getTimeRemainingRegularGame() {
+	public Double getTimeRemainingRegularGame() {
 		return component2.timeRemainingRegularGame;
 	}
 
-	public Short getTimeRemainingQuarter() {
+	public Double getTimeRemainingQuarter() {
 		return component2.timeRemainingQuarter;
 	}
